requirements.txt
google-genai
wave
Pillow
streamlit
python-dotenv
io